requirements.txt
asgiref==3.8.1
Django==4.2.19
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-link-header-pagination==0.2.0
drf-redesign==0.4.0
mysqlclient==2.2.7
pillow==11.1.0
Pygments==2.19.1
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
